+++ iTrainer/Models/SetsModel.swift
@@ -11,7 +11,7 @@ struct SetsModel: DataSetItemProtocol, Identifiable {
     var id: UUID
     var index: Int
     var reps: Int
-    var weight: Int
+    var weight: Float
     
     init(model: SetsModelDB) {
         self.id = model.id
@@ -20,7 +20,7 @@ struct SetsModel: DataSetItemProtocol, Identifiable {
         self.weight = model.weight
     }
     
-    init(id: UUID = UUID(), index: Int = 0, reps: Int, weight: Int) {
+    init(id: UUID = UUID(), index: Int = 0, reps: Int, weight: Float) {
         self.id = id
         self.index = index
         self.reps = reps
